Add Intermediate Components for Home and SignIn
Change the existing components for Home and SignIn to Patient Home and
SignIn.  The add intermediate components that navigate to Patient or
Standby components.
