+++ components/HomeComponent.js
@@ -1,57 +1,26 @@
 import React from 'react'
-import { StyleSheet, Text, View } from 'react-native'
-import { Button, Slider } from 'react-native-elements'
+import { ActivityIndicator, StatusBar, StyleSheet, View } from 'react-native'
 import { connect } from 'react-redux'
-import { removeTimers, setTimer, signoutUser } from '../redux/ActionCreators'
 
 const mapStateToProps = state => {
   return {
-    timer: state.timer
+    auth: state.auth
   }
 }
 
-const mapDispatchToProps = (dispatch) => (
-  {
-    removeTimers: () => dispatch(removeTimers()),
-    setTimer: interval => dispatch(setTimer(interval)),
-    signoutUser: () => dispatch(signoutUser())
-  }
-)
-
 class Home extends React.Component {
-  constructor (props) {
-    super(props)
-
-    this.state = {
-      interval: this.props.timer.interval
-    }
-
-    this.handleIntervalChange = this.handleIntervalChange.bind(this)
+  componentDidMount () {
+    this.props.auth.isPatient
+      ? this.props.navigation.navigate('PatientHome')
+      : this.props.navigation.navigate('StandbyHome')
   }
 
-  handleIntervalChange (value) {
-    this.props.removeTimers()
-    this.props.setTimer(this.state.interval)
-  }
-
+  // Render any loading content that you like here
   render () {
     return (
       <View style = { styles.container }>
-        <Text style = { styles.title }>Check-In Interval</Text>
-        <Slider
-          maximumValue = { 10000 }
-          minimumValue = { 1000 }
-          onSlidingComplete = { value => this.handleIntervalChange(value) }
-          onValueChange = { value => this.setState({ interval: value }) }
-          step = { 1000 }
-          style = { styles.slider }
-          value = { this.props.timer.interval }
-        />
-        <Text>{ this.state.interval }</Text>
-        <Button
-          onPress = { () => this.props.signoutUser() }
-          title = "Sign Out"
-        />
+        <ActivityIndicator />
+        <StatusBar barStyle="default" />
       </View>
     )
   }
@@ -65,15 +34,8 @@ const styles = StyleSheet.create(
       flex: 1,
       justifyContent: 'center',
       padding: 20
-    },
-    slider: {
-      width: 100
-    },
-    title: {
-      fontSize: 20,
-      fontWeight: 'bold'
     }
   }
 )
 
-export default connect(mapStateToProps, mapDispatchToProps)(Home)
+export default connect(mapStateToProps)(Home)
